pyproject.toml
[tool.poetry]
name = "geoip-geosite-list"
version = "0.1.0"
description = ""
authors = ["actusnileh <actusnileh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
gitpython = "^3.1.43"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
